+++ AnagoLeaderboard/AnagoLeaderboard/Controllers/PlayerController.cs
@@ -1,10 +1,8 @@
-﻿using AnagoLeaderboard.Database;
+﻿using AnagoLeaderboard.Models;
 using AnagoLeaderboard.Models.RequestParameters;
 using AnagoLeaderboard.Models.Results;
 using AnagoLeaderboard.Services;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using System.Linq.Expressions;
 
 namespace AnagoLeaderboard.Controllers
 {
@@ -13,13 +11,10 @@ namespace AnagoLeaderboard.Controllers
     public class PlayerController : ControllerBase
     {
         private readonly PlayerService _playerService;
-        LogComponent.LogClass _logClient;
 
-        public PlayerController(DatabaseContext databaseContext)
+        public PlayerController(PlayerService playerService)
         {
-            _playerService = new PlayerService(databaseContext);
-            _logClient = new LogComponent.LogClass(Anago.Config.clsAnagoConfig.GetConfig().LogServiceURI, "TafelvoetbalServer");
-
+            _playerService = playerService;
         }
 
         [HttpPost("player")]
@@ -33,31 +28,53 @@ namespace AnagoLeaderboard.Controllers
             {
                 string message = ex.Message;
                 if (ex.InnerException != null) message += " " + ex.InnerException.Message;
-                Log("CreatePlayer", message);
                 throw ex;
             }
         }
 
+        [HttpPatch("player/{id}")]
+        public async Task<ActionResult> UpdatePlayerName(string id, [FromBody] string newName)
+        {
+            await _playerService.UpdatePlayerName(id, newName);
+            return Ok();
+        }
+
         [HttpGet("players")]
-        public async Task<List<Player>> GetPlayers()
+        public async Task<List<DynamicRatingPlayer>> GetPlayers(bool activeOnly = false)
         {
             try
             {
-                return await _playerService.GetPlayers();
-            } catch (Exception ex)
+                return await _playerService.GetPlayers(activeOnly);
+            } 
+            catch (Exception ex)
             {
                 string message = ex.Message;
                 if (ex.InnerException != null) message += " " + ex.InnerException.Message;
-                Log("CreatePlayer", message);
                 throw ex;
             }
         }
 
         [HttpGet("player/{id}")]
-        public async Task<Player> GetPlayer(string id)
+        public async Task<DynamicRatingPlayer> GetPlayer(string id)
         {
             return await _playerService.GetPlayer(id);
         }
+
+        [HttpPatch("player/{id}/avatar")]
+        public async Task<ActionResult> UpdateAvatar(string id, [FromForm] UpdateAvatarForm newAvatar)
+        {
+            try
+            {
+                await _playerService.UpdateAvatar(id, newAvatar.Avatar);
+                return Ok();
+            }  catch (Exception ex)
+            {
+                string message = ex.Message;
+                if (ex.InnerException != null) message += " " + ex.InnerException.Message;
+                throw ex;
+            }
+        }
+
 
         [HttpGet("player/{id}/avatar")]
         public IActionResult GetAvatar(string id)
@@ -67,23 +84,58 @@ namespace AnagoLeaderboard.Controllers
 
         }
 
+        [HttpGet("player/{id}/games/page/{pageNumber}")]
+        public async Task<PlayerGamePage> GetPlayerGamesPage(string id, int pageNumber)
+        {
+            return await _playerService.GetPlayerGames(id, pageNumber);
+        }
+        
+        [HttpGet("player/{id}/games")]
+        public async Task<List<Game>> GetPlayerGames(string id)
+        {
+            return await _playerService.GetPlayerGames(id);
+        }
+          
+
         [HttpGet("player/one/avatar")]
         public async Task<IActionResult> GetPlayerOneAvatar()
         {
             var fileBytes = await _playerService.GetPlayerOneAvatar();
             return File(fileBytes, "image/jpeg");
         }
-        
-        [HttpDelete("players")]
-        public async Task<IActionResult> DeletePlayers()
+
+        [HttpGet("player/{id}/rank")]
+        public async Task<int> GetPlayerRank(string id)
         {
+            return await _playerService.GetPlayerRank(id);
+        }
+
+        [HttpDelete("players")]
+        public async Task<IActionResult> DeletePlayers(string password)
+        {
+            if (!password.Equals("deleteplayer" + DateTime.Now.Date.DayOfWeek + DateTime.Now.Date.Hour)) throw new Exception("Enter the password");
             await _playerService.DeletePlayers();
             return Ok();
         }
 
-        private void Log(string function, string message)
+        [HttpDelete("player/{id}")]
+        public async Task<IActionResult> DeletePlayer(string id)
         {
-            _logClient.Log("", "", "", function, message, DateTime.Now, 0);
+            await _playerService.DeletePlayer(id);
+            return Ok();
+        }
+
+        [HttpGet("player/{id}/stats")]
+        public async Task<PlayerStatistics> GetPlayerStats(string id)
+        {
+            var result = await _playerService.GetPlayerStatistics(id);
+            return result;
+        }
+        
+        [HttpPatch("player/{id}/active/{active}")]
+        public async Task UpdatePlayerActive(string id, bool active)
+        {
+            await _playerService.UpdatePlayerActive(id, active);
         }
 
     }
